Add coreutils module
