+++ geemap/common.py
@@ -24,123 +24,14 @@ import zipfile
 import ee
 import ipywidgets as widgets
 from ipytree import Node, Tree
-from typing import Union, List, Dict, Optional, Tuple, Any
+from typing import Union, List, Dict, Optional, Any
+
+from .coreutils import *
 
 try:
     from IPython.display import display, IFrame, Javascript
 except ImportError:
     pass
-
-
-def ee_initialize(
-    token_name="EARTHENGINE_TOKEN",
-    auth_mode=None,
-    service_account=False,
-    auth_args={},
-    user_agent_prefix="geemap",
-    project=None,
-    **kwargs,
-):
-    """Authenticates Earth Engine and initialize an Earth Engine session
-
-    Args:
-        token_name (str, optional): The name of the Earth Engine token.
-            Defaults to "EARTHENGINE_TOKEN". In Colab, you can also set a secret
-            named "EE_PROJECT_ID" to initialize Earth Engine.
-        auth_mode (str, optional): The authentication mode, can be one of colab,
-            notebook, localhost, or gcloud.
-            See https://developers.google.com/earth-engine/guides/auth for more
-            details. Defaults to None.
-        service_account (bool, optional): If True, use a service account.
-            Defaults to False.
-        auth_args (dict, optional): Additional authentication parameters for
-            aa.Authenticate(). Defaults to {}.
-        user_agent_prefix (str, optional): If set, the prefix (version-less)
-            value used for setting the user-agent string. Defaults to "geemap".
-        project (str, optional): The Google cloud project ID for Earth Engine.
-            Defaults to None.
-        kwargs (dict, optional): Additional parameters for ee.Initialize().
-            For example, opt_url='https://earthengine-highvolume.googleapis.com'
-            to use the Earth Engine High-Volume platform. Defaults to {}.
-    """
-    import httplib2
-    from .__init__ import __version__
-
-    user_agent = f"{user_agent_prefix}/{__version__}"
-    ee.data.setUserAgent(user_agent)
-
-    if "http_transport" not in kwargs:
-        kwargs["http_transport"] = httplib2.Http()
-
-    if project is None:
-        kwargs["project"] = get_api_key("EE_PROJECT_ID")
-    else:
-        kwargs["project"] = project
-
-    if auth_mode is None:
-        if in_colab_shell() and (ee.data._credentials is None):
-            ee.Authenticate()
-            ee.Initialize(**kwargs)
-            return
-        else:
-            auth_mode = "notebook"
-
-    auth_args["auth_mode"] = auth_mode
-
-    if ee.data._credentials is None:
-        ee_token = get_api_key(token_name)
-        if service_account:
-            try:
-                credential_file_path = os.path.expanduser(
-                    "~/.config/earthengine/private-key.json"
-                )
-
-                if os.path.exists(credential_file_path):
-                    with open(credential_file_path) as f:
-                        token_dict = json.load(f)
-                else:
-                    token_name = "EARTHENGINE_TOKEN"
-                    ee_token = os.environ.get(token_name)
-                    token_dict = json.loads(ee_token, strict=False)
-                service_account = token_dict["client_email"]
-                private_key = token_dict["private_key"]
-
-                credentials = ee.ServiceAccountCredentials(
-                    service_account, key_data=private_key
-                )
-                ee.Initialize(credentials, **kwargs)
-
-            except Exception as e:
-                raise Exception(e)
-
-        else:
-            try:
-                if ee_token is not None:
-                    credential_file_path = os.path.expanduser(
-                        "~/.config/earthengine/credentials"
-                    )
-                    if not os.path.exists(credential_file_path):
-                        os.makedirs(
-                            os.path.dirname(credential_file_path), exist_ok=True
-                        )
-                        if ee_token.startswith("{") and ee_token.endswith(
-                            "}"
-                        ):  # deals with token generated by new auth method (earthengine-api>=0.1.304).
-                            token_dict = json.loads(ee_token)
-                            with open(credential_file_path, "w") as f:
-                                f.write(json.dumps(token_dict))
-                        else:
-                            credential = (
-                                '{"refresh_token":"%s"}' % ee_token
-                            )  # deals with token generated by old auth method.
-                            with open(credential_file_path, "w") as f:
-                                f.write(credential)
-
-                ee.Initialize(**kwargs)
-
-            except Exception:
-                ee.Authenticate(**auth_args)
-                ee.Initialize(**kwargs)
 
 
 def ee_export_image(
@@ -1706,16 +1597,6 @@ def set_proxy(port=1080, ip="http://127.0.0.1", timeout=300):
         print(e)
 
 
-def in_colab_shell():
-    """Tests if the code is being executed within Google Colab."""
-    import sys
-
-    if "google.colab" in sys.modules:
-        return True
-    else:
-        return False
-
-
 def is_drive_mounted():
     """Checks whether Google Drive is mounted in Google Colab.
 
@@ -2045,23 +1926,6 @@ def is_tool(name):
     return shutil.which(name) is not None
 
 
-def random_string(string_length=3):
-    """Generates a random string of fixed length.
-
-    Args:
-        string_length (int, optional): Fixed length. Defaults to 3.
-
-    Returns:
-        str: A random string
-    """
-    import random
-    import string
-
-    # random.seed(1001)
-    letters = string.ascii_lowercase
-    return "".join(random.choice(letters) for i in range(string_length))
-
-
 def open_image_from_url(url, timeout=300, proxies=None):
     """Loads an image from the specified URL.
 
@@ -2215,85 +2079,6 @@ def upload_to_imgur(in_gif):
 ########################################
 #           Color and Fonts            #
 ########################################
-
-
-def rgb_to_hex(rgb=(255, 255, 255)):
-    """Converts RGB to hex color. In RGB color R stands for Red, G stands for Green, and B stands for Blue, and it ranges from the decimal value of 0 – 255.
-
-    Args:
-        rgb (tuple, optional): RGB color code as a tuple of (red, green, blue). Defaults to (255, 255, 255).
-
-    Returns:
-        str: hex color code
-    """
-    return "%02x%02x%02x" % rgb
-
-
-def hex_to_rgb(value="FFFFFF"):
-    """Converts hex color to RGB color.
-
-    Args:
-        value (str, optional): Hex color code as a string. Defaults to 'FFFFFF'.
-
-    Returns:
-        tuple: RGB color as a tuple.
-    """
-    value = value.lstrip("#")
-    lv = len(value)
-    return tuple(int(value[i : i + lv // 3], 16) for i in range(0, lv, lv // 3))
-
-
-def check_color(in_color):
-    """Checks the input color and returns the corresponding hex color code.
-
-    Args:
-        in_color (str or tuple): It can be a string (e.g., 'red', '#ffff00', 'ffff00', 'ff0') or RGB tuple (e.g., (255, 127, 0)).
-
-    Returns:
-        str: A hex color code.
-    """
-    import colour
-
-    out_color = "#000000"  # default black color
-    if isinstance(in_color, tuple) and len(in_color) == 3:
-        # rescale color if necessary
-        if all(isinstance(item, int) for item in in_color):
-            in_color = [c / 255.0 for c in in_color]
-
-        return colour.Color(rgb=tuple(in_color)).hex_l
-
-    else:
-        # try to guess the color system
-        try:
-            return colour.Color(in_color).hex_l
-
-        except Exception as e:
-            pass
-
-        # try again by adding an extra # (GEE handle hex codes without #)
-        try:
-            return colour.Color(f"#{in_color}").hex_l
-
-        except Exception as e:
-            print(
-                f"The provided color ({in_color}) is invalid. Using the default black color."
-            )
-            print(e)
-
-        return out_color
-
-
-def to_hex_colors(colors):
-    """Convert a GEE color palette into hexadecimal color codes. can handle mixin formats
-
-    Args:
-        colors (list): A list of hex color codes.
-
-    Returns:
-        list: A list of hex color codes prefixed with #.
-    """
-
-    return [check_color(c) for c in colors]
 
 
 def system_fonts(show_full_path=False):
@@ -2736,69 +2521,6 @@ def csv_to_vector(
     """
     gdf = csv_to_gdf(in_csv, latitude, longitude, encoding)
     gdf.to_file(output, **kwargs)
-
-
-def geojson_to_ee(geo_json, geodesic=False, encoding="utf-8"):
-    """Converts a geojson to ee.Geometry()
-
-    Args:
-        geo_json (dict): A geojson geometry dictionary or file path.
-        geodesic (bool, optional): Whether line segments should be interpreted as spherical geodesics. If false, indicates that line segments should be interpreted as planar lines in the specified CRS. If absent, defaults to true if the CRS is geographic (including the default EPSG:4326), or to false if the CRS is projected. Defaults to False.
-        encoding (str, optional): The encoding of characters. Defaults to "utf-8".
-
-    Returns:
-        ee_object: An ee.Geometry object
-    """
-
-    try:
-        if isinstance(geo_json, str):
-            if geo_json.startswith("http") and geo_json.endswith(".geojson"):
-                geo_json = github_raw_url(geo_json)
-                out_geojson = temp_file_path(extension=".geojson")
-                download_file(geo_json, out_geojson)
-                with open(out_geojson, "r", encoding=encoding) as f:
-                    geo_json = json.loads(f.read())
-                os.remove(out_geojson)
-
-            elif os.path.isfile(geo_json):
-                with open(os.path.abspath(geo_json), encoding=encoding) as f:
-                    geo_json = json.load(f)
-
-        # geo_json["geodesic"] = geodesic
-        if geo_json["type"] == "FeatureCollection":
-            for feature in geo_json["features"]:
-                if feature["geometry"]["type"] != "Point":
-                    feature["geometry"]["geodesic"] = geodesic
-            features = ee.FeatureCollection(geo_json)
-            return features
-        elif geo_json["type"] == "Feature":
-            geom = None
-            if "style" in geo_json["properties"]:
-                keys = geo_json["properties"]["style"].keys()
-                if "radius" in keys:  # Checks whether it is a circle
-                    geom = ee.Geometry(geo_json["geometry"])
-                    radius = geo_json["properties"]["style"]["radius"]
-                    geom = geom.buffer(radius)
-                elif geo_json["geometry"]["type"] == "Point":
-                    geom = ee.Geometry(geo_json["geometry"])
-                else:
-                    geom = ee.Geometry(geo_json["geometry"], "", geodesic)
-            elif (
-                geo_json["geometry"]["type"] == "Point"
-            ):  # Checks whether it is a point
-                coordinates = geo_json["geometry"]["coordinates"]
-                longitude = coordinates[0]
-                latitude = coordinates[1]
-                geom = ee.Geometry.Point(longitude, latitude)
-            else:
-                geom = ee.Geometry(geo_json["geometry"], "", geodesic)
-            return geom
-        else:
-            raise Exception("Could not convert the geojson to ee.Geometry()")
-
-    except Exception as e:
-        print("Could not convert the geojson to ee.Geometry()")
-        raise Exception(e)
 
 
 def ee_to_geojson(ee_object, filename=None, indent=2, **kwargs):
@@ -3652,18 +3374,6 @@ def screen_capture(filename, monitor=1):
 ########################################
 
 
-def open_url(url):
-    """Opens the URL in a new browser tab."""
-    if in_colab_shell():
-        display(
-            Javascript('window.open("{url}", "_blank", "noopener")'.format(url=url))
-        )
-    else:
-        import webbrowser
-
-        webbrowser.open_new_tab(url)
-
-
 def api_docs():
     """Open a browser and navigate to the geemap API documentation."""
     import webbrowser
@@ -4394,37 +4104,6 @@ def ee_data_html(asset):
 
     except Exception as e:
         print(e)
-
-
-def create_code_cell(code="", where="below"):
-    """Creates a code cell in the IPython Notebook.
-
-    Args:
-        code (str, optional): Code to fill the new code cell with. Defaults to ''.
-        where (str, optional): Where to add the new code cell. It can be one of the following: above, below, at_bottom. Defaults to 'below'.
-    """
-
-    import base64
-    import pyperclip
-
-    from IPython.display import Javascript, display
-
-    try:
-        pyperclip.copy(str(code))
-    except Exception as e:
-        pass
-
-    encoded_code = (base64.b64encode(str.encode(code))).decode()
-    display(
-        Javascript(
-            """
-        var code = IPython.notebook.insert_cell_{0}('code');
-        code.set_text(atob("{1}"));
-    """.format(
-                where, encoded_code
-            )
-        )
-    )
 
 
 def ee_api_to_csv(outfile=None, timeout=300, proxies=None):
@@ -8786,27 +8465,6 @@ def jpg_to_gif(in_dir, out_gif, fps=10, loop=0):
     )
 
 
-def geometry_type(ee_object):
-    """Get geometry type of an Earth Engine object.
-
-    Args:
-        ee_object (object): An Earth Engine object.
-
-    Returns:
-        str: Returns geometry type. One of Point, MultiPoint, LineString, LinearRing, MultiLineString, BBox, Rectangle, Polygon, MultiPolygon.
-    """
-    if isinstance(ee_object, ee.Geometry):
-        return ee_object.type().getInfo()
-    elif isinstance(ee_object, ee.Feature):
-        return ee_object.geometry().type().getInfo()
-    elif isinstance(ee_object, ee.FeatureCollection):
-        return ee.Feature(ee_object.first()).geometry().type().getInfo()
-    else:
-        raise TypeError(
-            "The ee_object must be one of ee.Geometry, ee.Feature, ee.FeatureCollection."
-        )
-
-
 def vector_styling(
     ee_object,
     column,
@@ -9688,34 +9346,6 @@ def osm_to_geojson(query, which_result=None, by_osmid=False, buffer_dist=None):
     """
     gdf = osm_to_gdf(query, which_result, by_osmid, buffer_dist)
     return gdf.__geo_interface__
-
-
-def get_api_key(name: Optional[str] = None, key: Optional[str] = None) -> Optional[str]:
-    """
-    Retrieves an API key. If a key is provided, it is returned directly. If a
-    name is provided, the function attempts to retrieve the key from user data
-    (if running in Google Colab) or from environment variables.
-
-    Args:
-        name (Optional[str], optional): The name of the key to retrieve. Defaults to None.
-        key (Optional[str], optional): The key to return directly. Defaults to None.
-
-    Returns:
-        Optional[str]: The retrieved key, or None if no key was found.
-    """
-
-    if key is not None:
-        return key
-    elif name is not None:
-        if in_colab_shell():
-            from google.colab import userdata
-
-            try:
-                return userdata.get(name)
-            except:
-                return os.environ.get(name)
-        else:
-            return os.environ.get(name)
 
 
 def planet_monthly_tropical(api_key=None, token_name="PLANET_API_KEY"):
@@ -10667,27 +10297,6 @@ def gdf_to_geojson(gdf, out_geojson=None, epsg=None):
             gdf.to_file(out_geojson, driver="GeoJSON")
     except Exception as e:
         raise Exception(e)
-
-
-def temp_file_path(extension):
-    """Returns a temporary file path.
-
-    Args:
-        extension (str): The file extension.
-
-    Returns:
-        str: The temporary file path.
-    """
-
-    import tempfile
-    import uuid
-
-    if not extension.startswith("."):
-        extension = "." + extension
-    file_id = str(uuid.uuid4())
-    file_path = os.path.join(tempfile.gettempdir(), f"{file_id}{extension}")
-
-    return file_path
 
 
 def get_temp_dir():
@@ -11924,71 +11533,6 @@ def write_lidar(source, destination, do_compress=None, laz_backend=None):
     source.write(destination, do_compress=do_compress, laz_backend=laz_backend)
 
 
-def download_file(
-    url=None,
-    output=None,
-    quiet=False,
-    proxy=None,
-    speed=None,
-    use_cookies=True,
-    verify=True,
-    id=None,
-    fuzzy=False,
-    resume=False,
-    unzip=True,
-    overwrite=False,
-):
-    """Download a file from URL, including Google Drive shared URL.
-
-    Args:
-        url (str, optional): Google Drive URL is also supported. Defaults to None.
-        output (str, optional): Output filename. Default is basename of URL.
-        quiet (bool, optional): Suppress terminal output. Default is False.
-        proxy (str, optional): Proxy. Defaults to None.
-        speed (float, optional): Download byte size per second (e.g., 256KB/s = 256 * 1024). Defaults to None.
-        use_cookies (bool, optional): Flag to use cookies. Defaults to True.
-        verify (bool | str, optional): Either a bool, in which case it controls whether the server's TLS certificate is verified, or a string, in which case it must be a path to a CA bundle to use. Default is True.. Defaults to True.
-        id (str, optional): Google Drive's file ID. Defaults to None.
-        fuzzy (bool, optional): Fuzzy extraction of Google Drive's file Id. Defaults to False.
-        resume (bool, optional): Resume the download from existing tmp file if possible. Defaults to False.
-        unzip (bool, optional): Unzip the file. Defaults to True.
-        overwrite (bool, optional): Overwrite the file if it already exists. Defaults to False.
-
-    Returns:
-        str: The output file path.
-    """
-
-    import gdown
-
-    if output is None:
-        if isinstance(url, str) and url.startswith("http"):
-            output = os.path.basename(url)
-
-    if isinstance(url, str):
-        if os.path.exists(os.path.abspath(output)) and (not overwrite):
-            print(
-                f"{output} already exists. Skip downloading. Set overwrite=True to overwrite."
-            )
-            return os.path.abspath(output)
-        else:
-            url = github_raw_url(url)
-
-    if "https://drive.google.com/file/d/" in url:
-        fuzzy = True
-
-    output = gdown.download(
-        url, output, quiet, proxy, speed, use_cookies, verify, id, fuzzy, resume
-    )
-
-    if unzip and output.endswith(".zip"):
-        with zipfile.ZipFile(output, "r") as zip_ref:
-            if not quiet:
-                print("Extracting files...")
-            zip_ref.extractall(os.path.dirname(output))
-
-    return os.path.abspath(output)
-
-
 def download_folder(
     url=None,
     id=None,
@@ -12651,38 +12195,6 @@ def image_count(
         image = image.clip(region)
 
     return image
-
-
-def check_cmap(cmap):
-    """Check the colormap and return a list of colors.
-
-    Args:
-        cmap (str | list | Box): The colormap to check.
-
-    Returns:
-        list: A list of colors.
-    """
-
-    from box import Box
-    from .colormaps import get_palette
-
-    if isinstance(cmap, str):
-        try:
-            palette = get_palette(cmap)
-            if isinstance(palette, dict):
-                palette = palette["default"]
-            return palette
-        except Exception as e:
-            try:
-                return check_color(cmap)
-            except Exception as e:
-                raise Exception(f"{cmap} is not a valid colormap.")
-    elif isinstance(cmap, Box):
-        return list(cmap["default"])
-    elif isinstance(cmap, list) or isinstance(cmap, tuple):
-        return cmap
-    else:
-        raise Exception(f"{cmap} is not a valid colormap.")
 
 
 def dynamic_world(
@@ -13558,21 +13070,6 @@ def setupJS():
         )
 
 
-def github_raw_url(url):
-    """Get the raw URL for a GitHub file.
-
-    Args:
-        url (str): The GitHub URL.
-    Returns:
-        str: The raw URL.
-    """
-    if isinstance(url, str) and url.startswith("https://github.com/") and "blob" in url:
-        url = url.replace("github.com", "raw.githubusercontent.com").replace(
-            "blob/", "", 1
-        )
-    return url
-
-
 def change_require(lib_path):
     if not isinstance(lib_path, str):
         raise ValueError("lib_path must be a string.")
@@ -14150,114 +13647,6 @@ def reproject(image, output, dst_crs="EPSG:4326", resampling="nearest", **kwargs
                     resampling=resampling,
                     **kwargs,
                 )
-
-
-def get_info(ee_object, layer_name="", opened=False, return_node=False):
-    """Print out the information for an Earth Engine object using a tree structure. The source code
-        was adapted from https://github.com/google/earthengine-jupyter. Credits to Tyler Erickson.
-
-    Args:
-        ee_object (object): The Earth Engine object.
-        layer_name (str, optional): The name of the layer. Defaults to "".
-        opened (bool, optional): Whether to expand the tree. Defaults to False.
-        return_node (bool, optional): Whether to return the widget as ipytree.Node.
-            If False, returns the widget as ipytree.Tree. Defaults to False.
-    """
-
-    def _order_items(item_dict, ordering_list):
-        """Orders dictionary items in a specified order.
-        Adapted from https://github.com/google/earthengine-jupyter.
-        """
-        list_of_tuples = [
-            (key, item_dict[key])
-            for key in [x for x in ordering_list if x in item_dict.keys()]
-        ]
-
-        return dict(list_of_tuples)
-
-    def _process_info(info):
-        node_list = []
-        if isinstance(info, list):
-            for count, item in enumerate(info):
-                if isinstance(item, (list, dict)):
-                    if "id" in item:
-                        count = f"{count}: \"{item['id']}\""
-                    if "data_type" in item:
-                        count = f"{count}, {item['data_type']['precision']}"
-                    if "crs" in item:
-                        count = f"{count}, {item['crs']}"
-                    if "dimensions" in item:
-                        dimensions = item["dimensions"]
-                        count = f"{count}, {dimensions[0]}x{dimensions[1]} px"
-                    node_list.append(
-                        Node(f"{count}", nodes=_process_info(item), opened=opened)
-                    )
-                else:
-                    node_list.append(Node(f"{count}: {item}", icon="file"))
-        elif isinstance(info, dict):
-            for k, v in info.items():
-                if isinstance(v, (list, dict)):
-                    if k == "properties":
-                        k = f"properties: Object ({len(v)} properties)"
-                    elif k == "bands":
-                        k = f"bands: List ({len(v)} elements)"
-                    node_list.append(
-                        Node(f"{k}", nodes=_process_info(v), opened=opened)
-                    )
-                else:
-                    node_list.append(Node(f"{k}: {v}", icon="file"))
-        else:
-            node_list.append(Node(f"{info}", icon="file"))
-        return node_list
-
-    if isinstance(ee_object, ee.FeatureCollection):
-        ee_object = ee_object.map(lambda f: ee.Feature(None, f.toDictionary()))
-    layer_info = ee_object.getInfo()
-    if not layer_info:
-        return None
-
-    props = layer_info.get("properties", {})
-    layer_info["properties"] = dict(sorted(props.items()))
-
-    ordering_list = []
-    if "type" in layer_info:
-        ordering_list.append("type")
-        ee_type = layer_info["type"]
-    else:
-        ee_type = ""
-
-    if "id" in layer_info:
-        ordering_list.append("id")
-        ee_id = layer_info["id"]
-    else:
-        ee_id = ""
-
-    ordering_list.append("version")
-    ordering_list.append("bands")
-    ordering_list.append("properties")
-
-    layer_info = _order_items(layer_info, ordering_list)
-    nodes = _process_info(layer_info)
-
-    if len(layer_name) > 0:
-        layer_name = f"{layer_name}: "
-
-    if "bands" in layer_info:
-        band_info = f' ({len(layer_info["bands"])} bands)'
-    else:
-        band_info = ""
-    root_node = Node(f"{layer_name}{ee_type} {band_info}", nodes=nodes, opened=opened)
-    # root_node.open_icon = "plus-square"
-    # root_node.open_icon_style = "success"
-    # root_node.close_icon = "minus-square"
-    # root_node.close_icon_style = "info"
-
-    if return_node:
-        return root_node
-    else:
-        tree = Tree()
-        tree.add_node(root_node)
-        return tree
 
 
 def download_3dep_lidar(region, filename, scale=1.0, crs="EPSG:3857"):
@@ -15654,133 +15043,6 @@ def get_ee_token():
         return None
 
 
-def widget_template(
-    widget=None,
-    opened=True,
-    show_close_button=True,
-    widget_icon="gear",
-    close_button_icon="times",
-    widget_args={},
-    close_button_args={},
-    display_widget=None,
-    m=None,
-    position="topright",
-):
-    """Create a widget template.
-
-    Args:
-        widget (ipywidgets.Widget, optional): The widget to be displayed. Defaults to None.
-        opened (bool, optional): Whether to open the toolbar. Defaults to True.
-        show_close_button (bool, optional): Whether to show the close button. Defaults to True.
-        widget_icon (str, optional): The icon name for the toolbar button. Defaults to 'gear'.
-        close_button_icon (str, optional): The icon name for the close button. Defaults to "times".
-        widget_args (dict, optional): Additional arguments to pass to the toolbar button. Defaults to {}.
-        close_button_args (dict, optional): Additional arguments to pass to the close button. Defaults to {}.
-        display_widget (ipywidgets.Widget, optional): The widget to be displayed when the toolbar is clicked.
-        m (geemap.Map, optional): The geemap.Map instance. Defaults to None.
-        position (str, optional): The position of the toolbar. Defaults to "topright".
-    """
-
-    name = "_" + random_string()  # a random attribute name
-
-    if "value" not in widget_args:
-        widget_args["value"] = False
-    if "tooltip" not in widget_args:
-        widget_args["tooltip"] = "Toolbar"
-    if "layout" not in widget_args:
-        widget_args["layout"] = widgets.Layout(
-            width="28px", height="28px", padding="0px 0px 0px 4px"
-        )
-    widget_args["icon"] = widget_icon
-
-    if "value" not in close_button_args:
-        close_button_args["value"] = False
-    if "tooltip" not in close_button_args:
-        close_button_args["tooltip"] = "Close the tool"
-    if "button_style" not in close_button_args:
-        close_button_args["button_style"] = "primary"
-    if "layout" not in close_button_args:
-        close_button_args["layout"] = widgets.Layout(
-            height="28px", width="28px", padding="0px 0px 0px 4px"
-        )
-    close_button_args["icon"] = close_button_icon
-
-    try:
-        toolbar_button = widgets.ToggleButton(**widget_args)
-    except:
-        widget_args.pop("layout")
-        toolbar_button = widgets.ToggleButton(**widget_args)
-        toolbar_button.layout.width = "28px"
-        toolbar_button.layout.height = "28px"
-        toolbar_button.layout.padding = "0px 0px 0px 4px"
-
-    try:
-        close_button = widgets.ToggleButton(**close_button_args)
-    except:
-        close_button_args.pop("layout")
-        close_button = widgets.ToggleButton(**close_button_args)
-        close_button.layout.width = "28px"
-        close_button.layout.height = "28px"
-        close_button.layout.padding = "0px 0px 0px 4px"
-
-    toolbar_widget = widgets.VBox()
-    toolbar_widget.children = [toolbar_button]
-    toolbar_header = widgets.HBox()
-    if show_close_button:
-        toolbar_header.children = [close_button, toolbar_button]
-    else:
-        toolbar_header.children = [toolbar_button]
-    toolbar_footer = widgets.VBox()
-
-    if widget is not None:
-        toolbar_footer.children = [
-            widget,
-        ]
-    else:
-        toolbar_footer.children = []
-
-    def toolbar_btn_click(change):
-        if change["new"]:
-            close_button.value = False
-            toolbar_widget.children = [toolbar_header, toolbar_footer]
-            if display_widget is not None:
-                widget.outputs = ()
-                with widget:
-                    display(display_widget)
-        else:
-            toolbar_widget.children = [toolbar_button]
-
-    toolbar_button.observe(toolbar_btn_click, "value")
-
-    def close_btn_click(change):
-        if change["new"]:
-            toolbar_button.value = False
-            if m is not None:
-                control = getattr(m, name)
-                if control is not None and control in m.controls:
-                    m.remove_control(control)
-                    delattr(m, name)
-            toolbar_widget.close()
-
-    close_button.observe(close_btn_click, "value")
-
-    toolbar_button.value = opened
-    if m is not None:
-        import ipyleaflet
-
-        toolbar_control = ipyleaflet.WidgetControl(
-            widget=toolbar_widget, position=position
-        )
-
-        if toolbar_control not in m.controls:
-            m.add_control(toolbar_control)
-
-            setattr(m, name, toolbar_control)
-
-    else:
-        return toolbar_widget
-
-
 def geotiff_to_image(image: str, output: str) -> None:
     """
     Converts a GeoTIFF file to a JPEG/PNG image.
@@ -16213,35 +15475,6 @@ def is_on_aws():
         if item.endswith(".aws") or "ec2-user" in item:
             on_aws = True
     return on_aws
-
-
-def _get_colab_secret(key: str) -> Optional[str]:
-    """Returns a Colab secret (if available), otherwise None."""
-    if in_colab_shell():
-        from google.colab import userdata
-
-        try:
-            return userdata.get(key)
-        except (userdata.SecretNotFoundError, userdata.NotebookAccessError):
-            return None  # Secret doesn't exist or insufficient access.
-    return None
-
-
-def get_google_maps_api_key(key: str = "GOOGLE_MAPS_API_KEY") -> Optional[str]:
-    """
-    Retrieves the Google Maps API key from the environment or Colab user data.
-
-    Args:
-        key (str, optional): The name of the environment variable or Colab user
-            data key where the API key is stored. Defaults to
-            'GOOGLE_MAPS_API_KEY'.
-
-    Returns:
-        str: The API key, or None if it could not be found.
-    """
-    if api_key := _get_colab_secret(key):
-        return api_key
-    return os.environ.get(key, None)
 
 
 def xarray_to_raster(dataset, filename: str, **kwargs: Dict[str, Any]) -> None:
